voice-prescription-fastapi-main/requirements.txt
nltk
gunicorn
spacy
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.2.5/en_ner_bc5cdr_md-0.2.5.tar.gz
torchvision
transformers
numpy
seqeval
fastapi
uvicorn
pandas

ipython
filelock
pydantic
allennlp
scikit-learn
networkx
matplotlib
seaborn
